Rename to xapple and refine constants
- Renamed `apple` package to `xapple`
- Adjusted constant names accordingly (e.g. `PropertyCalendarColor` instead of `PropertyXAppleCalendarColor` given the package context).
