+++ extensions/wg/wg.go
@@ -4,17 +4,17 @@ import (
 	ical "github.com/arran4/golang-ical"
 )
 
-// PropertyXWgTimezone is the X-WG-TIMEZONE property
-const PropertyXWgTimezone = "X-WG-TIMEZONE"
+// PropertyTimezone is the X-WG-TIMEZONE property
+const PropertyTimezone = "X-WG-TIMEZONE"
 
-// PropertyXWgCalName is the X-WG-CALNAME property
-const PropertyXWgCalName = "X-WG-CALNAME"
+// PropertyCalName is the X-WG-CALNAME property
+const PropertyCalName = "X-WG-CALNAME"
 
-// PropertyXWgCalDesc is the X-WG-CALDESC property
-const PropertyXWgCalDesc = "X-WG-CALDESC"
+// PropertyCalDesc is the X-WG-CALDESC property
+const PropertyCalDesc = "X-WG-CALDESC"
 
-// PropertyXWgRelCalId is the X-WG-RELCALID property
-const PropertyXWgRelCalId = "X-WG-RELCALID"
+// PropertyRelCalId is the X-WG-RELCALID property
+const PropertyRelCalId = "X-WG-RELCALID"
 
 // SetProperty allows extending the properties easily
 func SetProperty(cal *ical.Calendar, property string, value string, params ...ical.PropertyParameter) {
@@ -23,20 +23,20 @@ func SetProperty(cal *ical.Calendar, property string, value string, params ...ic
 
 // SetTimezone sets the X-WG-TIMEZONE property
 func SetTimezone(cal *ical.Calendar, tz string, params ...ical.PropertyParameter) {
-	SetProperty(cal, PropertyXWgTimezone, tz, params...)
+	SetProperty(cal, PropertyTimezone, tz, params...)
 }
 
 // SetCalName sets the X-WG-CALNAME property
 func SetCalName(cal *ical.Calendar, name string, params ...ical.PropertyParameter) {
-	SetProperty(cal, PropertyXWgCalName, name, params...)
+	SetProperty(cal, PropertyCalName, name, params...)
 }
 
 // SetCalDesc sets the X-WG-CALDESC property
 func SetCalDesc(cal *ical.Calendar, desc string, params ...ical.PropertyParameter) {
-	SetProperty(cal, PropertyXWgCalDesc, desc, params...)
+	SetProperty(cal, PropertyCalDesc, desc, params...)
 }
 
 // SetRelCalId sets the X-WG-RELCALID property
 func SetRelCalId(cal *ical.Calendar, id string, params ...ical.PropertyParameter) {
-	SetProperty(cal, PropertyXWgRelCalId, id, params...)
+	SetProperty(cal, PropertyRelCalId, id, params...)
 }
